refactor: 역할 구분을 위해 Lotto 클래스의 일부 메소드를 LottoMachine으로 이동
1. Lotto 클래스는 당첨번호만 가진다
2. LottoMachine 클래스에서 Lotto의 인스턴스와 bonusNumber를 가진다.
3. Lotto의 인스턴스와 bonusNumber는 외부에서 set 해준다.
4. 따라서 LottoMachine 내부에서 통계를 구할 수 있다.
