+++ src/LottoMachine.js
@@ -1,16 +1,35 @@
-const { RESULT } = require('./utils/constants');
+const { RESULT, LOTTO_PRICE } = require('./utils/constants');
 const Lotto = require('./Lotto');
+const { getAscending } = require('./utils/common');
+const { Random } = require('@woowacourse/mission-utils');
 
 class LottoMachine {
   generatedLottos = [];
+  lotto;
+  bonusNumber;
 
   constructor() {}
 
-  generateLotto(amount) {
-    const number = amount / 1000;
+  setLottoWinningNumbers(lottoWinningNumbers) {
+    this.lotto = new Lotto(lottoWinningNumbers);
+  }
+
+  setBonusNumber(bonusNumber) {
+    this.bonusNumber = bonusNumber;
+  }
+
+  #generate() {
+    let numbers = Random.pickUniqueNumbersInRange(1, 45, 6);
+    numbers = getAscending(numbers);
+
+    return numbers;
+  }
+
+  generateLotto(purchasedAmount) {
+    const number = purchasedAmount / LOTTO_PRICE;
 
     for (let i = 0; i < number; i++) {
-      const lotto = Lotto.generate();
+      const lotto = this.#generate();
       this.generatedLottos.push(lotto);
     }
   }
@@ -23,7 +42,29 @@ class LottoMachine {
     return this.generatedLottos.length;
   }
 
-  getStatistic(lotto, bonusNumber) {
+  #getResult(generatedLotto) {
+    let matchCount = 0;
+
+    generatedLotto.forEach((number) => {
+      if (this.lotto.getWinningNumbers().includes(number)) matchCount += 1;
+    });
+
+    switch (matchCount) {
+      case 6:
+        return RESULT.FIRST;
+      case 5:
+        if (generatedLotto.includes(this.bonusNumber)) return RESULT.SECOND;
+        return RESULT.THIRD;
+      case 4:
+        return RESULT.FOURTH;
+      case 3:
+        return RESULT.FIFTH;
+      default:
+        return RESULT.LAST;
+    }
+  }
+
+  getStatistic() {
     const statistic = {
       FIRST: 0,
       SECOND: 0,
@@ -33,7 +74,7 @@ class LottoMachine {
     };
 
     for (const generatedLotto of this.generatedLottos) {
-      const result = lotto.getResult(generatedLotto, bonusNumber);
+      const result = this.#getResult(generatedLotto);
 
       switch (result) {
         case RESULT.FIRST:
